Fix slither and solhint issues

+++ contracts/token/erc721/preset/ImmutableERC721MintByID.sol
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache 2.0
 pragma solidity 0.8.19;
 
-import "../abstract/ImmutableERC721Base.sol";
+import {ImmutableERC721Base} from "../abstract/ImmutableERC721Base.sol";
 
 contract ImmutableERC721MintByID is ImmutableERC721Base {
     ///     =====   Constructor  =====
@@ -29,9 +29,17 @@ contract ImmutableERC721MintByID is ImmutableERC721Base {
         address royaltyReceiver_,
         uint96 feeNumerator_
     )
-        ImmutableERC721Base(owner_, name_, symbol_, baseURI_, contractURI_, operatorAllowlist_, royaltyReceiver_, feeNumerator_)
+        ImmutableERC721Base(
+            owner_,
+            name_,
+            symbol_,
+            baseURI_,
+            contractURI_,
+            operatorAllowlist_,
+            royaltyReceiver_,
+            feeNumerator_
+        )
     {}
-
 
     /** @notice Allows minter to mint `tokenID` to `to`
      *  @param to the address to mint the token to
@@ -76,7 +84,7 @@ contract ImmutableERC721MintByID is ImmutableERC721Base {
      * @param tokenIDs an array of token IDs to burn
      */
     function burnBatch(uint256[] calldata tokenIDs) external {
-        for (uint i = 0; i < tokenIDs.length; i++) {
+        for (uint256 i = 0; i < tokenIDs.length; i++) {
             burn(tokenIDs[i]);
         }
     }
@@ -98,7 +106,7 @@ contract ImmutableERC721MintByID is ImmutableERC721Base {
             revert IImmutableERC721MismatchedTransferLengths();
         }
 
-        for (uint i = 0; i < tr.tokenIds.length; i++) {
+        for (uint256 i = 0; i < tr.tokenIds.length; i++) {
             safeTransferFrom(tr.from, tr.tos[i], tr.tokenIds[i]);
         }
     }
